updated yosys file for comp, comp_nbit_ncc is not supported yet

+++ Benchmarks2/compare/compile.sh
@@ -1,8 +1,16 @@
-#!/bin/bash
+echo "Use -d to compile with Synopsys Design Compiler"
+
 mkdir -p syn
 
-design_vision -no_gui -f comp.dcsh
-rm *.pvl *.syn *.mr *.log *.svf
+if [ $1 ] && [ $1 = "-d" ]; then
+	design_vision -no_gui -f comp.dcsh
+	rm *.pvl *.syn *.mr *.log *.svf
+else
+	yosys -c comp.tcl
+fi
 
-../../Verilog2EMP/bin/V2EMP_Main -i syn/aes_11cc_syn.v -o syn/aes_11cc_syn.emp --log2std
-../../Verilog2EMP/bin/V2EMP_Main -i syn/aes_11cc_syn.v -o syn/aes_11cc_syn.emp --log2std
+for verilogfile in syn/*.v
+do
+  empfile=${verilogfile%.*}.emp
+  ../../Verilog2EMP/bin/V2EMP_Main -i $verilogfile -o $empfile --log2std 
+done
